Recalculate field widths after condition changes
